+++ src/Symfony/Component/Form/FormConfig.php
@@ -54,6 +54,11 @@ class FormConfig implements FormConfigEditorInterface
     private $virtual = false;
 
     /**
+     * @var Boolean
+     */
+    private $compound = false;
+
+    /**
      * @var array
      */
     private $types = array();
@@ -112,6 +117,11 @@ class FormConfig implements FormConfigEditorInterface
      * @var string
      */
     private $dataClass;
+
+    /**
+     * @var Boolean
+     */
+    private $dataLocked;
 
     /**
      * @var array
@@ -359,6 +369,14 @@ class FormConfig implements FormConfigEditorInterface
     /**
      * {@inheritdoc}
      */
+    public function getCompound()
+    {
+        return $this->compound;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
     public function getTypes()
     {
         return $this->types;
@@ -492,6 +510,14 @@ class FormConfig implements FormConfigEditorInterface
     public function getDataClass()
     {
         return $this->dataClass;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getDataLocked()
+    {
+        return $this->dataLocked;
     }
 
     /**
@@ -635,6 +661,16 @@ class FormConfig implements FormConfigEditorInterface
     /**
      * {@inheritdoc}
      */
+    public function setCompound($compound)
+    {
+        $this->compound = $compound;
+
+        return $this;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
     public function setTypes(array $types)
     {
         $this->types = $types;
@@ -653,6 +689,16 @@ class FormConfig implements FormConfigEditorInterface
     }
 
     /**
+     * {@inheritdoc}
+     */
+    public function setDataLocked($locked)
+    {
+        $this->dataLocked = $locked;
+
+        return $this;
+    }
+
+    /**
      * Validates whether the given variable is a valid form name.
      *
      * @param string $name The tested form name.
@@ -660,7 +706,7 @@ class FormConfig implements FormConfigEditorInterface
      * @throws UnexpectedTypeException   If the name is not a string.
      * @throws \InvalidArgumentException If the name contains invalid characters.
      */
-    static public function validateName($name)
+    public static function validateName($name)
     {
         if (!is_string($name)) {
             throw new UnexpectedTypeException($name, 'string');
@@ -688,7 +734,7 @@ class FormConfig implements FormConfigEditorInterface
      *
      * @return Boolean Whether the name is valid.
      */
-    static public function isValidName($name)
+    public static function isValidName($name)
     {
         return '' === $name || preg_match('/^[a-zA-Z0-9_][a-zA-Z0-9_\-:]*$/D', $name);
     }