+++ src/Symfony/Component/Form/Tests/Extension/Core/Type/DateTypeTest.php
@@ -12,7 +12,6 @@
 namespace Symfony\Component\Form\Tests\Extension\Core\Type;
 
 use Symfony\Component\Form\Extension\Core\View\ChoiceView;
-use Symfony\Component\Form\FormError;
 
 class DateTypeTest extends LocalizedTestCase
 {
@@ -251,37 +250,45 @@ class DateTypeTest extends LocalizedTestCase
     /**
      * This test is to check that the strings '0', '1', '2', '3' are no accepted
      * as valid IntlDateFormatter constants for FULL, LONG, MEDIUM or SHORT respectively.
+     *
+     * @expectedException Symfony\Component\OptionsResolver\Exception\InvalidOptionsException
      */
-    public function testFormatOptionCustomPatternCollapsingIntlDateFormatterConstant()
+    public function testThrowExceptionIfFormatIsNoPattern()
     {
-        $form = $this->factory->create('date', null, array(
+        $this->factory->create('date', null, array(
             'format' => '0',
             'widget' => 'single_text',
             'input' => 'string',
         ));
-
-        $form->setData('2010-06-02');
-
-        // This would be what would be outputed if '0' was mistaken for \IntlDateFormatter::FULL
-        $this->assertNotEquals('Mittwoch, 02. Juni 2010', $form->getViewData());
     }
 
     /**
-     * @expectedException Symfony\Component\Form\Exception\CreationException
+     * @expectedException Symfony\Component\OptionsResolver\Exception\InvalidOptionsException
      */
-    public function testValidateFormatOptionGivenWrongConstants()
+    public function testThrowExceptionIfFormatDoesNotContainYearMonthAndDay()
     {
-        $form = $this->factory->create('date', null, array(
+        $this->factory->create('date', null, array(
+            'months' => array(6, 7),
+            'format' => 'yy',
+        ));
+    }
+
+    /**
+     * @expectedException Symfony\Component\OptionsResolver\Exception\InvalidOptionsException
+     */
+    public function testThrowExceptionIfFormatIsNoConstant()
+    {
+        $this->factory->create('date', null, array(
             'format' => 105,
         ));
     }
 
     /**
-     * @expectedException Symfony\Component\Form\Exception\CreationException
+     * @expectedException Symfony\Component\OptionsResolver\Exception\InvalidOptionsException
      */
-    public function testValidateFormatOptionGivenArrayValue()
+    public function testThrowExceptionIfFormatIsInvalid()
     {
-        $form = $this->factory->create('date', null, array(
+        $this->factory->create('date', null, array(
             'format' => array(),
         ));
     }
@@ -335,21 +342,6 @@ class DateTypeTest extends LocalizedTestCase
     {
         $form = $this->factory->create('date', null, array(
             'months' => array(6, 7),
-        ));
-
-        $view = $form->createView();
-
-        $this->assertEquals(array(
-            new ChoiceView('6', '06'),
-            new ChoiceView('7', '07'),
-        ), $view->get('month')->getVar('choices'));
-    }
-
-    public function testMonthsOptionNumericIfFormatContainsNoMonth()
-    {
-        $form = $this->factory->create('date', null, array(
-            'months' => array(6, 7),
-            'format' => 'yy',
         ));
 
         $view = $form->createView();
@@ -546,5 +538,89 @@ class DateTypeTest extends LocalizedTestCase
 
         $view = $form->createView();
         $this->assertEquals('date', $view->getVar('type'));
+    }
+
+    public function testPassDefaultEmptyValueToViewIfNotRequired()
+    {
+        $form = $this->factory->create('date', null, array(
+            'required' => false,
+        ));
+
+        $view = $form->createView();
+        $this->assertSame('', $view->get('year')->getVar('empty_value'));
+        $this->assertSame('', $view->get('month')->getVar('empty_value'));
+        $this->assertSame('', $view->get('day')->getVar('empty_value'));
+    }
+
+    public function testPassNoEmptyValueToViewIfRequired()
+    {
+        $form = $this->factory->create('date', null, array(
+            'required' => true,
+        ));
+
+        $view = $form->createView();
+        $this->assertNull($view->get('year')->getVar('empty_value'));
+        $this->assertNull($view->get('month')->getVar('empty_value'));
+        $this->assertNull($view->get('day')->getVar('empty_value'));
+    }
+
+    public function testPassEmptyValueAsString()
+    {
+        $form = $this->factory->create('date', null, array(
+            'empty_value' => 'Empty',
+        ));
+
+        $view = $form->createView();
+        $this->assertSame('Empty', $view->get('year')->getVar('empty_value'));
+        $this->assertSame('Empty', $view->get('month')->getVar('empty_value'));
+        $this->assertSame('Empty', $view->get('day')->getVar('empty_value'));
+    }
+
+    public function testPassEmptyValueAsArray()
+    {
+        $form = $this->factory->create('date', null, array(
+            'empty_value' => array(
+                'year' => 'Empty year',
+                'month' => 'Empty month',
+                'day' => 'Empty day',
+            ),
+        ));
+
+        $view = $form->createView();
+        $this->assertSame('Empty year', $view->get('year')->getVar('empty_value'));
+        $this->assertSame('Empty month', $view->get('month')->getVar('empty_value'));
+        $this->assertSame('Empty day', $view->get('day')->getVar('empty_value'));
+    }
+
+    public function testPassEmptyValueAsPartialArray_addEmptyIfNotRequired()
+    {
+        $form = $this->factory->create('date', null, array(
+            'required' => false,
+            'empty_value' => array(
+                'year' => 'Empty year',
+                'day' => 'Empty day',
+            ),
+        ));
+
+        $view = $form->createView();
+        $this->assertSame('Empty year', $view->get('year')->getVar('empty_value'));
+        $this->assertSame('', $view->get('month')->getVar('empty_value'));
+        $this->assertSame('Empty day', $view->get('day')->getVar('empty_value'));
+    }
+
+    public function testPassEmptyValueAsPartialArray_addNullIfRequired()
+    {
+        $form = $this->factory->create('date', null, array(
+            'required' => true,
+            'empty_value' => array(
+                'year' => 'Empty year',
+                'day' => 'Empty day',
+            ),
+        ));
+
+        $view = $form->createView();
+        $this->assertSame('Empty year', $view->get('year')->getVar('empty_value'));
+        $this->assertNull($view->get('month')->getVar('empty_value'));
+        $this->assertSame('Empty day', $view->get('day')->getVar('empty_value'));
     }
 }