+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -32,3 +32,9 @@ CHANGELOG
   `gc_probability`/`gc_divisor` chance of being run. The `gc_maxlifetime` defines
    how long a session can idle for. It is different from cookie lifetime which
    declares how long a cookie can be stored on the remote client.
+ * Removed 'auto_start' configuration parameter from session config. The session will
+   start on demand.
+ * [BC BREAK] TemplateNameParser::parseFromFilename() has been moved to a dedicated
+   parser: TemplateFilenameParser::parse().
+ * [BC BREAK] Kernel parameters are replaced by their value whereever they appear
+   in Route patterns, requirements and defaults. Use '%%' as the escaped value for '%'.